+++ app/src/types/self.d.ts
@@ -1,11 +1,3 @@
-interface Dated {
-  date: string;
-}
-
-interface Heighted {
-  height: number;
-}
-
 interface Valued {
   value: number;
 }