+++ parser/src/io/consts.rs
@@ -1,2 +1,2 @@
-pub const IMPORTS_FOLDER_PATH: &str = "./imports";
+pub const INPUTS_FOLDER_PATH: &str = "./in";
 pub const OUTPUTS_FOLDER_PATH: &str = "./target/outputs";