+++ parser/src/price/kraken.rs
@@ -62,7 +62,7 @@ impl Kraken {
                     .collect::<BTreeMap<_, _>>())
             },
             10,
-            5,
+            10,
         )
     }
 
@@ -117,7 +117,7 @@ impl Kraken {
                     .collect::<BTreeMap<_, _>>())
             },
             10,
-            5,
+            10,
         )
     }
 }