+++ server/run.sh
@@ -1,5 +1,5 @@
 if cargo watch --help &> /dev/null; then
-    cargo watch --no-vcs-ignores -w "./src" -w "./run.sh" -w ".trigger" -x "run -r"
+    cargo watch --no-vcs-ignores -w "./src" -w "./run.sh" -w "./in/datasets_len.txt" -x "run -r"
 else
     cargo run -r
 fi